requirements.txt
requests==2.31.0
pyyaml==6.0.1
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
